# docs/contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Song Record API
  version: 0.7.1
  description: P07-01 공통 계약. implemented=기존 인증 API, planned=후속 구현 목표. 문서의 planned 경로는 현재 서버에서 제공되지 않는다. 인증 DTO는
    호환성을 위해 camelCase, 업무 DTO는 snake_case를 사용한다. 문자열 길이는 정규화 후 Unicode code point 수이며 메모는 CRLF/CR을 LF로 통일한다.
servers:
- url: /v1
security:
- BearerAuth: []
paths:
...
  /conditions:
    get:
      operationId: listConditions
      summary: listConditions
      description: 계정별 기본/사용자 컨디션 목록. code를 녹음의 condition_code에 사용한다. 기본 항목은 기존 VERY_GOOD/GOOD/NORMAL/BAD
        코드, 사용자 항목은 UUID 코드다. state=ACTIVE(기본)/ARCHIVED/ALL 및 정규화 이름·UUID 커서 정렬을 지원한다.
      x-implementation-status: implemented
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - name: state
        in: query
        schema:
          type: string
          enum:
          - ACTIVE
          - ARCHIVED
          - ALL
          default: ACTIVE
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConditionPage'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
    post:
      operationId: createCondition
      summary: createCondition
      description: 계정별 컨디션 생성. 사용자 정의 code는 id와 같은 UUID이며 직접 지정할 수 없다. 이름은 공백 제거 후 1~50
        code point, 활성 정규화 이름 중복은 409. 기본 컨디션도 계정별로 수정/보관할 수 있다.
      x-implementation-status: implemented
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Condition'
        '201':
          description: 생성
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Condition'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConditionCreate'
  /conditions/{id}:
    patch:
      operationId: renameCondition
      summary: renameCondition
      description: revision을 검사하여 활성 컨디션 이름만 수정한다. 과거 recording_tag.name_snapshot은 변경하지
        않는다. 이름 중복/보관 컨디션/버전 충돌은 409.
      x-implementation-status: implemented
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Condition'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConditionPatch'
  /conditions/{id}/archive:
    post:
      operationId: archiveCondition
      summary: archiveCondition
      description: 컨디션를 물리 삭제하지 않고 archived_at을 기록한다. 기존 녹음 관계와 당시 이름은 유지하며 신규 선택은 차단한다.
        이미 보관된 컨디션의 최초 보관 시각은 유지한다.
      x-implementation-status: implemented
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/DeviceId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Condition'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConditionArchive'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque session token
  parameters:
    DeviceId:
      name: X-Device-Id
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
      description: 로그인 응답에서 받은 현재 계정의 등록 기기 UUID.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
      description: 업무 변경 op_id. 같은 키·동일 정규화 본문은 결과 재사용. 다른 본문은 409 IDEMPOTENCY_CONFLICT.
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      schema:
        $ref: '#/components/schemas/OpaqueCursor'
    RequestId:
      name: X-Request-Id
      in: header
      schema:
        $ref: '#/components/schemas/RequestId'
      description: 생략하거나 유효하지 않으면 서버가 새 추적 ID를 발급한다.
  schemas:
    Uuid:
      type: string
      format: uuid
      pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
    UtcInstant:
      type: string
      format: date-time
      pattern: ^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d{1,3})?Z$
      description: UTC 시각. 최대 밀리초 정밀도. 초 단위 인증 응답도 허용한다.
...
    Revision:
      type: integer
      format: int64
      minimum: 1
      maximum: 9223372036854775807
...
    ConditionCode:
      type: string
      enum:
      - VERY_GOOD
      - GOOD
      - NORMAL
      - BAD
...
    OpaqueCursor:
      type: string
      minLength: 1
      maxLength: 8192
      description: 내용을 해석하거나 직접 구성하지 않는다. 계정·조회 종류·필터·정렬·키 규격·limit에 결합한다. 불일치나 변조는 400 INVALID_CURSOR, 조회 세대 변경 또는 만료는 409 LIST_CURSOR_EXPIRED로 첫 페이지 재조회를 요구한다. 목록 커서 수명은 발급 후 30분이며 동기화 스냅샷을 뜻하지 않는다.
...
    RequestId:
      type: string
      pattern: ^[A-Za-z0-9._-]{8,64}$
...
    ApiError:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          pattern: ^[A-Z][A-Z0-9_]*$
        message:
          type: string
          minLength: 1
        retryable:
          type: boolean
        request_id:
          $ref: '#/components/schemas/RequestId'
        details:
          type: object
          additionalProperties: true
          description: 허용된 구조화 값만. 토큰·파일 URL·타 계정 정보·메모 원문 금지.
      required:
      - code
      - message
      - retryable
      - request_id
      - details
    ApiErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
      required:
      - error
...
    Condition:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          $ref: '#/components/schemas/ConditionName'
        revision:
          $ref: '#/components/schemas/Revision'
        archived_at:
          anyOf:
          - $ref: '#/components/schemas/UtcInstant'
          - type: 'null'
        updated_at:
          $ref: '#/components/schemas/UtcInstant'
        code:
          $ref: '#/components/schemas/ConditionReference'
      required:
      - id
      - name
      - revision
      - archived_at
      - updated_at
      - code
    ConditionCreate:
      type: object
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          $ref: '#/components/schemas/ConditionName'
      required:
      - id
      - name
    ConditionPatch:
      type: object
      additionalProperties: false
      properties:
        base_revision:
          $ref: '#/components/schemas/Revision'
        name:
          $ref: '#/components/schemas/ConditionName'
      required:
      - base_revision
      - name
    ConditionArchive:
      type: object
      additionalProperties: false
      properties:
        base_revision:
          $ref: '#/components/schemas/Revision'
      required:
      - base_revision
    ConditionPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Condition'
        count:
          type: integer
          minimum: 0
        next_cursor:
          anyOf:
          - $ref: '#/components/schemas/OpaqueCursor'
          - type: 'null'
      required:
      - items
      - count
      - next_cursor
    ConditionReference:
      oneOf:
      - $ref: '#/components/schemas/ConditionCode'
      - $ref: '#/components/schemas/Uuid'
      description: 계정별 기본 컨디션의 호환 코드 또는 사용자 컨디션 UUID. API 목록의 code 값을 선택/필터에 사용한다.
    ConditionName:
      type: string
      minLength: 1
      maxLength: 50
  responses:
    Error400:
      description: 공통 오류 응답.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Error401:
      description: 공통 오류 응답.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    Error404:
      description: 리소스가 없거나 현재 계정 소유가 아님. 두 경우를 구분하지 않는다.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Error409:
      description: 상태·revision·멱등 키 충돌.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    Error500:
      description: 공통 오류 응답.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Error503:
      description: 공통 오류 응답.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'